+++ ChangeLog.md
@@ -9,6 +9,7 @@
 
 * packages REST API calls implemented.
 * unit test `test_login_failed_invalid_url` disabled because it delays all tests.
+* have unit tests for packages.
 
 ## V1.5.1